Added tools files for external functions
- moved and modified fillDeck from Pile class to tools.cpp
- added getDeckSize function to ask for user input
- added distributeCards function to distribute cards to players' decks
- fixed warnings

+++ Player.h
@@ -31,11 +31,11 @@ public:
     Card playCard();
 
     // Compter les points
-    int computePoints();
+    double computePoints();
 
 private:
     std::string m_name;
-    int totalPoints;
+    double totalPoints;
     Pile<Card> m_deck, m_gain;
 };
